+++ Add Code Here/C/Average Number.c/Average Number.c
