--- design.md
@@ -1,5 +1,0 @@
-## System Design
-
-- LLD
-- HLD
-- SOLID Principles
